backend/our_groceries/requirements.txt
django==2.2.7
djangorestframework==3.10.3
drf-yasg==1.17.0